Updated AzureApiService
- Moved CreateApiClient and GetAccessTokenSync methods to separate
  files.
- Created CreateArmClient() method.
- Added UpdateApiClient() method for generating a new access token when
  API calls are made. Has been added to GetHostPools() and
  GetSessionHosts() methods.
- Added logging.
