+++ services/azure/desktop-virtualization/AzureApiService_GetSessionHosts.cs
@@ -29,11 +29,15 @@ public partial class AzureApiService : IAzureApiService
     /// <returns>An array of all sessions hosts currently in the hostpool.</returns>
     private async Task<List<SessionHost>?> GetSessionHostsAsync(AvdHostPool hostPool)
     {
+        // Refresh the API client with a new token.
+        UpdateApiClient();
+
         HttpRequestMessage requestMessage = new(
             method: HttpMethod.Get,
             requestUri: $"{hostPool.HostPoolResourceId}/sessionHosts?api-version=2021-07-12"
         );
 
+        logger.LogInformation("Sending API call to '{RequestUri}'", requestMessage.RequestUri);
         HttpResponseMessage responseMessage = await apiClient.SendAsync(requestMessage);
 
         List<SessionHost>? sessionHosts = null;
